perf(message): replace vim.b.messages with lua variable (#898)
## Description

Using `vim.b` variables seems slow for big tables. This PR improves
performance of command output redirection (and possibly other places).
From my tests this is an order of 20 sec -> 4 sec.

Also, it looks like the current code sometimes assumes that
`vim.b[buf].messages` is a list of buffers, while it is a list of
message ids. I also changed it from a list to a set, because it's mostly
used for random access.

## Related Issue(s)

This helps in case described here:
https://github.com/folke/noice.nvim/issues/332 (filetype is still
applied many times, but from my recent benchmarks it seems that
`vim.b.messages` had bigger impact).