fix(hover): handle hover resize. fully Fixes #1122
Previous fix for #1122 didn't handle the case where the hover window was
still larger than the available space to display it. This fixes that
case as well as correctly aligns hover when displaying to the left.

Also use vim.go.lines instead of vim.fn.winheight to allow popups to
overlap other windows. That seems like a better choice since it's a
momentary popup vs making the popup potentially very small